sam/functions/pdw_downloader/requirements.txt
snowflake-connector-python==2.9.0
snowflake-sqlalchemy==1.4.7
pyarrow==8.0.0
pandas==1.5.2
aws-lambda-powertools[tracer]==2.15.0